requirements.txt
flask
bcrypt
python-dotenv
regex